requirements.txt
Flask==2.2.5
Flask-Session==0.4.0
Flask-Mail==0.9.1
email-validator==1.3.0
itsdangerous==2.1.2
Werkzeug==2.2.3
bcrypt==4.0.1
SQLAlchemy==1.4.50
Flask-SQLAlchemy==2.5.1
python-dotenv==1.0.0


#sqlite3-bro==0.1.1  # on verra si j'ai une bdd

